requirements.txt
dnspython==2.3.0
numpy==1.24.2
pandas==2.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
